cleanup and remove unused dependencies
