+++ src/main/groovy/io/beapi/api/config/BeapiWebAutoConfiguration.groovy
@@ -124,12 +124,12 @@ public class BeapiWebAutoConfiguration implements WebMvcConfigurer, BeanFactoryA
 		return messageSource;
 	}
 
-	@Bean
-	public LocalValidatorFactoryBean getValidator() {
-		LocalValidatorFactoryBean bean = new LocalValidatorFactoryBean();
-		bean.setValidationMessageSource(messageSource());
-		return bean;
-	}
+	//@Bean
+	//public LocalValidatorFactoryBean getValidator() {
+	//	LocalValidatorFactoryBean bean = new LocalValidatorFactoryBean();
+	//	bean.setValidationMessageSource(messageSource());
+	//	return bean;
+	//}
 
 
 	@Bean
