+++ src/main/groovy/io/beapi/api/service/BatchExchangeService.groovy
@@ -16,15 +16,12 @@
  */
 package io.beapi.api.service
 
-import io.beapi.api.service.ApiExchange
 import org.json.JSONObject
-import io.beapi.api.utils.ErrorCodes
 import org.slf4j.LoggerFactory
 import org.springframework.stereotype.Service
 import org.springframework.context.ApplicationContext
 import javax.json.*
 import org.springframework.security.web.header.*
-import groovyx.gpars.*
 import javax.servlet.http.HttpServletRequest
 import javax.servlet.http.HttpServletResponse
 
@@ -33,7 +30,7 @@ import javax.servlet.http.HttpServletResponse
 @Service
 public class BatchExchangeService extends ApiExchange{
 
-	private static final org.slf4j.Logger logger = LoggerFactory.getLogger(BatchExchangeService.class);
+	//private static final org.slf4j.Logger logger = LoggerFactory.getLogger(BatchExchangeService.class);
 	LinkedHashMap returns = [:]
 	LinkedHashMap networkRoles
 	LinkedList batch = []
@@ -51,20 +48,11 @@ public class BatchExchangeService extends ApiExchange{
 	}
 
 	boolean batchRequest(HttpServletRequest request, HttpServletResponse response, String authority) {
-		def post = request.getAttribute('POST')
-		def get = request.getAttribute('GET')
-
-		if(post['batchVars']){
-			if(!request.getAttribute('batchVars')) { request.setAttribute('batchVars', post['batchVars']) }
-			post.remove('batchVars')
-		}
 
 		initVars(request,response,authority)
 
-		LinkedHashMap<String,String> output = get + post
-		request.setAttribute('params',output)
-
 		setBatchParams(request)
+
 		if(!validateMethod()){
 			writeErrorResponse(response,'405',request.getRequestURI());
 		}
@@ -233,37 +221,4 @@ public class BatchExchangeService extends ApiExchange{
 		}
 	}
 
-	// Todo : Move to exchangeService??
-	/**
-	 * Standardized error handler for all interceptors; simplifies RESPONSE error handling in interceptors
-	 * @param HttpServletResponse response
-	 * @param String statusCode
-	 * @return LinkedHashMap commonly formatted linkedhashmap
-	 */
-	void writeErrorResponse(HttpServletResponse response, String statusCode, String uri){
-
-		response.setContentType("application/json")
-		response.setStatus(Integer.valueOf(statusCode))
-		String message = "{\"timestamp\":\"${System.currentTimeMillis()}\",\"status\":\"${statusCode}\",\"error\":\"${ErrorCodes.codes[statusCode]['short']}\",\"message\": \"${ErrorCodes.codes[statusCode]['long']}\",\"path\":\"${uri}\"}"
-		response.getWriter().write(message)
-		response.writer.flush()
-	}
-
-	// Todo : Move to exchangeService??
-	/**
-	 * Standardized error handler for all interceptors; simplifies RESPONSE error handling in interceptors
-	 * @param HttpServletResponse response
-	 * @param String statusCode
-	 * @return LinkedHashMap commonly formatted linkedhashmap
-	 */
-	void writeErrorResponse(HttpServletResponse response, String statusCode, String uri, String msg){
-		response.setContentType("application/json")
-		response.setStatus(Integer.valueOf(statusCode))
-		if(msg.isEmpty()){
-			msg = ErrorCodes.codes[statusCode]['long']
-		}
-		String message = "{\"timestamp\":\"${System.currentTimeMillis()}\",\"status\":\"${statusCode}\",\"error\":\"${ErrorCodes.codes[statusCode]['short']}\",\"message\": \"${msg}\",\"path\":\"${uri}\"}"
-		response.getWriter().write(message)
-		response.writer.flush()
-	}
 }
