+++ src/main/groovy/io/beapi/api/service/TraceExchangeService.groovy
@@ -17,14 +17,10 @@
 package io.beapi.api.service
 
 
-import io.beapi.api.service.ApiExchange
-import io.beapi.api.utils.ErrorCodes
-import io.beapi.api.service.TraceService
 import org.slf4j.LoggerFactory
 import org.springframework.stereotype.Service
 import javax.json.*
 import org.springframework.security.web.header.*
-import groovyx.gpars.*
 import javax.servlet.http.HttpServletRequest
 import javax.servlet.http.HttpServletResponse
 
@@ -33,7 +29,7 @@ import javax.servlet.http.HttpServletResponse
 @Service
 public class TraceExchangeService extends ApiExchange{
 
-	private static final org.slf4j.Logger logger = LoggerFactory.getLogger(TraceExchangeService.class);
+	//private static final org.slf4j.Logger logger = LoggerFactory.getLogger(TraceExchangeService.class);
 	private static final ArrayList RESERVED_PARAM_NAMES = ['batch','chain']
 	String cacheHash
 	ApiCacheService apiCacheService
@@ -143,36 +139,4 @@ public class TraceExchangeService extends ApiExchange{
 		}
 	}
 
-	// Todo : Move to exchangeService??
-	/**
-	 * Standardized error handler for all interceptors; simplifies RESPONSE error handling in interceptors
-	 * @param HttpServletResponse response
-	 * @param String statusCode
-	 * @return LinkedHashMap commonly formatted linkedhashmap
-	 */
-	void writeErrorResponse(HttpServletResponse response, String statusCode, String uri){
-		response.setContentType("application/json")
-		response.setStatus(Integer.valueOf(statusCode))
-		String message = "{\"timestamp\":\"${System.currentTimeMillis()}\",\"status\":\"${statusCode}\",\"error\":\"${ErrorCodes.codes[statusCode]['short']}\",\"message\": \"${ErrorCodes.codes[statusCode]['long']}\",\"path\":\"${uri}\"}"
-		response.getWriter().write(message)
-		response.writer.flush()
-	}
-
-	// Todo : Move to exchangeService??
-	/**
-	 * Standardized error handler for all interceptors; simplifies RESPONSE error handling in interceptors
-	 * @param HttpServletResponse response
-	 * @param String statusCode
-	 * @return LinkedHashMap commonly formatted linkedhashmap
-	 */
-	void writeErrorResponse(HttpServletResponse response, String statusCode, String uri, String msg){
-		response.setContentType("application/json")
-		response.setStatus(Integer.valueOf(statusCode))
-		if(msg.isEmpty()){
-			msg = ErrorCodes.codes[statusCode]['long']
-		}
-		String message = "{\"timestamp\":\"${System.currentTimeMillis()}\",\"status\":\"${statusCode}\",\"error\":\"${ErrorCodes.codes[statusCode]['short']}\",\"message\": \"${msg}\",\"path\":\"${uri}\"}"
-		response.getWriter().write(message)
-		response.writer.flush()
-	}
 }
